+++ reactor-aeron-tools/src/main/java/reactor/aeron/demo/RateReporter.java
@@ -85,7 +85,7 @@ public class RateReporter implements Runnable, Disposable {
       final long totalFragments,
       final long totalBytes) {
     System.out.format(
-        "%.06g msgs/sec, %.03g MB/sec, totals %d messages %d MB payloads%n",
+        "%.10g msgs/sec, %.10g MB/sec, totals %d messages %d MB payloads%n",
         messagesPerSec, bytesPerSec / (1024 * 1024), totalFragments, totalBytes / (1024 * 1024));
   }
